requirements.txt
psycopg2-binary
uuid
python-dotenv

fastapi
uvicorn
pydantic

pytest